manifest 수정, sp->dp, create 화면 내 TextField, submit button 수정
