--- app/src/main/java/com/sainfe/todowithspot/TodayView.java
@@ -1,49 +1,0 @@
-package com.sainfe.todowithspot;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import java.util.ArrayList;
-
-public class TodayView extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_today_view);
-
-        ArrayList<Item> list = new ArrayList<>();
-
-        for (int i=0; i<100; i++) {
-            list.add(new Item("text "+ i)) ;
-            // TODO : 데이테베이스에서 전체 조회로 구현 예정
-        }
-
-        // 리사이클러뷰에 LinearLayoutManager 객체 지정.
-        RecyclerView recyclerView = findViewById(R.id.recycler1) ;
-        recyclerView.setLayoutManager(new LinearLayoutManager(this)) ;
-
-        // 리사이클러뷰에 SimpleTextAdapter 객체 지정.
-        SimpleTextAdapter adapter = new SimpleTextAdapter(list) ;
-        recyclerView.setAdapter(adapter) ;
-
-        //createButton 클릭 시 , create view 전환
-        View createButton = findViewById(R.id.create_todo_list_button);
-        createButton.setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View view) {
-                Intent createViewintent = new Intent(getApplicationContext(), CreateView.class);
-                startActivity(createViewintent);
-            }
-        });
-    }
-
-
-}
