--- app/src/main/res/drawable/input_field.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="350dp"
-    android:height="48dp"
-    android:viewportWidth="350"
-    android:viewportHeight="48">
-  <path
-      android:pathData="M8,0L342,0A8,8 0,0 1,350 8L350,40A8,8 0,0 1,342 48L8,48A8,8 0,0 1,0 40L0,8A8,8 0,0 1,8 0z"
-      android:fillColor="#F3F5F7"/>
-</vector>
