+++ app/src/main/res/layout/activity_create_view.xml
@@ -5,13 +5,14 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".CreateView">
+    android:background="@color/white"
+    tools:context=".create.CreateView">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
-        android:padding="16sp">
+        android:padding="16dp">
 
         <ImageButton
             android:id="@+id/backward_button"
@@ -31,20 +32,19 @@
             android:gravity="center"
             android:text="@string/create"
             android:textColor="@color/black"
-            android:textSize="24sp"
+            android:textSize="24dp"
             android:textStyle="bold" />
 
-        <TextView
+        <Button
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
             android:layout_weight="3"
-            android:clickable="true"
-            android:focusable="true"
             android:gravity="center"
             android:text="@string/submit"
             android:textColor="@color/black"
-            android:textSize="12sp" />
+            android:textSize="12dp"
+            android:background="@null" />
 
     </LinearLayout>
 
@@ -52,24 +52,21 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:padding="16sp">
+        android:padding="16dp">
 
         <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:clickable="false"
-            android:padding="8sp"
+            android:padding="8dp"
             android:text="@string/context"
             android:textColor="@color/black"
-            android:textSize="20sp"
+            android:textSize="20dp"
             android:textStyle="bold" />
 
-        <ImageButton
+        <com.yourssu.design.system.atom.SimpleTextField
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@null"
-            android:padding="8sp"
-            android:src="@drawable/input_field" />
+            android:layout_height="wrap_content" />
 
     </LinearLayout>
 
@@ -77,7 +74,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:padding="16sp">
+        android:padding="16dp">
 
         <TextView
             android:layout_width="match_parent"
@@ -86,7 +83,7 @@
             android:padding="8dp"
             android:text="@string/dateAndHour"
             android:textColor="@color/black"
-            android:textSize="20sp"
+            android:textSize="20dp"
             android:textStyle="bold" />
 
         <LinearLayout
@@ -125,7 +122,7 @@
             android:padding="8dp"
             android:text="@string/place"
             android:textColor="@color/black"
-            android:textSize="20sp"
+            android:textSize="20dp"
             android:textStyle="bold" />
 
         <LinearLayout
@@ -144,7 +141,7 @@
                 android:text="@string/departPlace"
                 android:textAlignment="center"
                 android:textColor="@color/black"
-                android:textSize="15sp" />
+                android:textSize="15dp" />
 
             <RadioButton
                 android:layout_width="match_parent"
@@ -157,7 +154,7 @@
                 android:text="@string/arrivePlace"
                 android:textAlignment="center"
                 android:textColor="@color/black"
-                android:textSize="15sp" />
+                android:textSize="15dp" />
 
         </LinearLayout>
     </LinearLayout>
