+++ app/src/main/res/layout/recyclerview_item.xml
@@ -10,8 +10,11 @@
         android:layout_height="wrap_content"
         android:clickable="true"
         android:focusable="false"
-        android:paddingTop="16sp"
-        android:paddingBottom="16sp"
+        android:paddingTop="16dp"
+        android:paddingBottom="16dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"
+        android:buttonTint="@color/black"
         android:textColor="@color/black"
         android:textSize="15sp"
         tools:ignore="MissingConstraints" />
